fix: item layout
